RFPGenie_github/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
sqlmodel
psycopg2-binary
supabase
google-generativeai
google-ads
sentence-transformers
python-multipart
litellm
openai
pydantic-settings
asyncpg
google-adk
greenlet
langchain-google-genai
pypdf
python-docx
python-dotenv